% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoFR_binary_iid_fun.R
\name{SoFR_binary_iid_fun}
\alias{SoFR_binary_iid_fun}
\title{Binary conditional model with functional covariate}
\usage{
SoFR_binary_iid_fun(
  t,
  X,
  y,
  nbd_index = NULL,
  alphaHat,
  betaHat,
  alpha_tol = 1e-08,
  beta_tol = 1e-06,
  test_beta = TRUE
)
}
\arguments{
\item{t}{A T-dimensional vector of containing densely time grid points, where the functions are evaluated.}

\item{X}{An n by T matrix of functional covariates. Each row represents one observed functional covariate.}

\item{y}{An n-dimensional vector of response variable.}

\item{nbd_index}{Not used in this function. Must be left as \code{NULL}.}

\item{alphaHat}{A numeric value specifying the initial values of \eqn{\alpha}.}

\item{betaHat}{A T-dimensional vector specifying the initial value of \eqn{\beta}.}

\item{alpha_tol}{Convergence tolerance for \eqn{\hat{\alpha}}, with a default of 1e-8.}

\item{beta_tol}{Convergence tolerance for \eqn{\hat{\beta}}, with a default of 1e-6.}

\item{test_beta}{Logical. 
If TRUE, performs a Wald test for the null hypothesis \eqn{\beta = 0}; otherwise, no test is performed.}
}
\value{
A list with the following elements:
\describe{
   \itemize{
      \item \code{iter}: The number of iterations.
      \item \code{p}: The selected truncation level.
      \item \code{alphaHat}: The estimated value of \eqn{\alpha}.
      \item \code{betaHat}: A T-dimensional vector of estimated parameter function \eqn{\hat{\beta}}.
      \item \code{pval}: p-value of Wald test for testing the null hypothesis \eqn{\beta=0}.
   }
}
}
\description{
Performs estimation using an iteratively reweighted least squares (IRLS) algorithm 
with spectral decomposition based on the working covariate.
}
\examples{

library(SoFRCond)

# Load example data included in the package
data(binary_data)
t <- binary_data$t; X <- binary_data$X; y <- binary_data$y; nbd_index <- binary_data$nbd_index
betaHat <- rep(0, length(t))

res_iid <- SoFR_binary_iid_fun(t, X, y, nbd_index = NULL, alphaHat = 0, betaHat = betaHat, test_beta = FALSE)

# Example: plot the estimated parameter function
plot(t, res_iid$betaHat, type = 'l')

}
\seealso{
\code{\link{SoFR_conti_iid_fun}}
\code{\link{SoFR_CAR_spa_fun}}
\code{\link{SoFR_binary_spa_fun}}
}
