\name{SoFRCond-package}
\alias{SoFRCond-package}
\alias{SoFRCond}
\docType{package}
\title{
\packageTitle{SoFRCond}
}
\description{
\packageDescription{SoFRCond}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{SoFRCond}

Maintainer: \packageMaintainer{SoFRCond}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = SoFRCond)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{SoFRCond}}
%\section{Documentation Index}{\packageIndices{SoFRCond}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
