% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoFR_conti_iid_fun.R
\name{SoFR_conti_iid_fun}
\alias{SoFR_conti_iid_fun}
\title{Linear model with functional covariate}
\usage{
SoFR_conti_iid_fun(t, X, y, num_nei, sigma2Hat = 1)
}
\arguments{
\item{t}{A T-dimensional vector of containing densely time grid points, where the functions are evaluated.}

\item{X}{An n by T matrix of functional covariates. Each row represents one observed functional covariate.}

\item{y}{An n-dimensional vector of response variable.}

\item{num_nei}{An n-dimensional vector specifying the number of neighbors for each spatial unit. 
Each element corresponds to the count of adjacent locations for a given spatial unit.
This is required because the conditional variance is defined as 
\eqn{\sigma_i^2 = \sigma^2 / w_{i+}}, 
where \eqn{w_{i+}} denotes the number of neighbors of spatial location \eqn{s_i}.}

\item{sigma2Hat}{A numeric value specifying an initial of \eqn{\sigma^2}, with a default of 1.}
}
\value{
A list with the following elements:
\describe{
   \itemize{
      \item \code{p}: The selected truncation level.
      \item \code{sigma2Hat}: The estimated value of \eqn{\sigma_i^2}.
      \item \code{betaHat}: A T-dimensional vector of estimated parameter function \eqn{\hat{\beta}}.
      \item \code{pval_beta}: P-value for testing the null hypothesis \eqn{\beta=0}.
   }
}
}
\description{
Performs estimation algorithm based on spectral decomposition of covariance operator of the functional covariate.
}
\examples{

library(SpatialSoFR)

# Load example data included in the package
data(CAR_data)
t <- CAR_data$t; X <- CAR_data$X; y <- CAR_data$y; nbd_index <- CAR_data$nbd_index
num_neighbors <- sapply(nbd_index, length)

res_iid <- SoFR_conti_iid_fun(t, X, y, num_neighbors)

# Example: plot the estimated parameter function
plot(t, res_iid$betaHat, type = 'l')

}
\seealso{
\code{\link{SoFR_CAR_spa_fun}}
\code{\link{SoFR_binary_iid_fun}}
\code{\link{SoFR_binary_spa_fun}}
}
